+++ package/CHANGELOG.md
@@ -28,7 +28,7 @@ use std::path::PathBuf;
 use filego::split::{
     Split, 
     SplitResult, 
-+   tokio::AsyncSplitExt as _,
++   tokio::SplitAsyncExt as _,
 };
 
 async fn example() {